changes in code

+++ JobOrbit/src/main/java/com/joborbit/JobOrbit/repository/UsersTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersTypeRepository extends JpaRepository<UsersType, Long> {
+public interface UsersTypeRepository extends JpaRepository<UsersType, Integer> {
 }
